+++ pyctr/type/nand.py
@@ -12,6 +12,8 @@ from threading import Lock
 from typing import NamedTuple, TYPE_CHECKING
 from weakref import WeakSet
 
+from pyfatfs.PyFatFS import PyFatBytesIOFS
+
 from ..common import PyCTRError
 from ..crypto import CryptoEngine, Keyslot
 from ..fileio import SubsectionIO
@@ -337,7 +339,7 @@ class NAND(TypeReaderCryptoBase):
 
     __slots__ = (
         '_base_files', '_lock', '_subfile', 'bonus_partitions', 'counter', 'counter_twl', 'ctr_index', 'ctr_partitions',
-        'essential', 'header', 'twl_index', 'twl_partitions'
+        'essential', 'header', 'twl_index', 'twl_partitions', '_fat_partitons'
     )
 
     essential: 'Optional[ExeFSReader]'
@@ -364,6 +366,9 @@ class NAND(TypeReaderCryptoBase):
         # opened files to close if the NAND is closed
         # noinspection PyTypeChecker
         self._open_files: Set[SubsectionIO] = WeakSet()
+        # FAT partitions should be closed before the underlying file so it can do some final cleanup
+        # noinspection PyTypeChecker
+        self._fat_partitons: Set[PyFatBytesIOFS] = WeakSet()
 
         self._file.seek(0, 2)
         raw_nand_size = self._file.tell() - self._start
@@ -598,6 +603,8 @@ class NAND(TypeReaderCryptoBase):
 
     def close(self):
         if not self.closed:
+            for f in self._fat_partitons:
+                f.close()
             for c in self._base_files.values():
                 c.close()
             for s in self._open_files:
@@ -606,7 +613,8 @@ class NAND(TypeReaderCryptoBase):
             if self.essential:
                 self.essential.close()
 
-            self._open_files = set()
+            self._fat_partitons = frozenset()
+            self._open_files = frozenset()
             self._base_files = {}
 
             super().close()
@@ -655,6 +663,21 @@ class NAND(TypeReaderCryptoBase):
         self._open_files.add(fh)
         return fh
 
+    def open_ctr_fat(self, partition_index: int = 0) -> 'PyFatBytesIOFS':
+        """
+        Opens a FAT filesystem inside a CTR partition for reading and writing.
+
+        In practice there is only ever one, so this opens it by default.
+
+        :param partition_index: Partition index number.
+        :return: A FAT filesystem.
+        :rtype: PyFatBytesIOFS
+        """
+        fh = self.open_ctr_partition(partition_index)
+        fat = PyFatBytesIOFS(fp=fh)
+        self._fat_partitons.add(fat)
+        return fat
+
     def open_twl_partition(self, partition_index: int):
         """
         Opens a raw partition in TWLNAND for reading and writing.
@@ -673,6 +696,21 @@ class NAND(TypeReaderCryptoBase):
         self._open_files.add(fh)
         return fh
 
+    def open_twl_fat(self, partition_index) -> 'PyFatBytesIOFS':
+        """
+        Opens a FAT filesystem inside a TWL partition for reading and writing.
+
+        0 is TWL NAND and 1 is TWL Photo.
+
+        :param partition_index: Partition index number.
+        :return: A FAT filesystem.
+        :rtype: PyFatBytesIOFS
+        """
+        fh = self.open_twl_partition(partition_index)
+        fat = PyFatBytesIOFS(fp=fh)
+        self._fat_partitons.add(fat)
+        return fat
+
     def open_bonus_partition(self):
         """
         Opens the GodMode9 bonus partition.
@@ -686,6 +724,21 @@ class NAND(TypeReaderCryptoBase):
         fh = SubsectionIO(self._subfile, bonus_info.offset + bonus_mbr_info[0], bonus_mbr_info[1])
         self._open_files.add(fh)
         return fh
+
+    def open_bonus_fat(self) -> 'PyFatBytesIOFS':
+        """
+        Opens the GodMode9 bonus FAT partition.
+
+        0 is TWL NAND and 1 is TWL Photo.
+
+        :param partition_index: Partition index number.
+        :return: A FAT filesystem.
+        :rtype: PyFatBytesIOFS
+        """
+        fh = self.open_bonus_partition()
+        fat = PyFatBytesIOFS(fp=fh)
+        self._fat_partitons.add(fat)
+        return fat
 
     def open_raw_section(self, section: 'Union[NANDSection, int]'):
         """